# FNV 1a 64-bit hash algorithm optimizations