--- a/CP77.Common/Tools/FNV1A/FNV1A64HashAlgorithm.cs
+++ b/CP77.Common/Tools/FNV1A/FNV1A64HashAlgorithm.cs
@@ -1,12 +1,13 @@
 ﻿using System;
+using System.Buffers;
 using System.Security.Cryptography;
 using System.Text;
 
-/// <summary>
-/// Hash algorithm, Fowler–Noll–Vo
-/// </summary>
 namespace CP77.Common.Tools.FNV1A
 {
+    /// <summary>
+    /// Hash algorithm, Fowler–Noll–Vo
+    /// </summary>
     public sealed class FNV1A64HashAlgorithm : HashAlgorithm
     {
         private const ulong FnvHashPrime = 0x00000100000001B3;
@@ -15,23 +16,13 @@ namespace CP77.Common.Tools.FNV1A
         private ulong fnvhash;
         private Encoding encoding;
 
-        public ulong HashUInt64
-        {
-            get
-            {
-                if (fnvhash == FnvHashInitial)
-                {
-                    return 0;
-                }
-                return fnvhash;
-            }
-        }
+        public ulong HashUInt64 => fnvhash == FnvHashInitial ? 0 : fnvhash;
 
         public FNV1A64HashAlgorithm()
             : this(Encoding.ASCII)
         {
-
         }
+
         public FNV1A64HashAlgorithm(Encoding encoding)
         {
             this.Initialize();
@@ -48,41 +39,29 @@ namespace CP77.Common.Tools.FNV1A
         {
             AppendString(value, false);
         }
+
         public void AppendString(string value, bool nullEnded)
         {
-            if (String.IsNullOrEmpty(value))
+            if (string.IsNullOrEmpty(value))
                 return;
 
-            var buffer = new char[nullEnded ? value.Length + 1 : value.Length];
-            value.CopyTo(0, buffer, 0, value.Length);
+            var length = encoding.GetMaxByteCount(nullEnded ? value.Length + 1 : value.Length);
+            var buffer = ArrayPool<byte>.Shared.Rent(length);
+            try
+            {
+                var encodedLength = encoding.GetBytes(value.AsSpan(), buffer.AsSpan());
+                if (nullEnded)
+                {
+                    buffer[encodedLength] = 0;
+                    encodedLength++;
+                }
 
-            var data = encoding.GetBytes(buffer);
-
-            HashCore(data, 0, data.Length);
-        }
-
-        public static ulong HashString(string value)
-        {
-            return HashString(value, Encoding.ASCII, false);
-        }
-        public static ulong HashString(string value, Encoding encoding)
-        {
-            return HashString(value, encoding, false);
-        }
-        public static ulong HashString(string value, Encoding encoding, bool nullEnded)
-        {
-            if (String.IsNullOrEmpty(value))
-                return 0;
-
-            var buffer = new char[nullEnded ? value.Length + 1 : value.Length];
-            value.CopyTo(0, buffer, 0, value.Length);
-
-            var fnv1a = new FNV1A64HashAlgorithm();
-            var data = encoding.GetBytes(buffer);
-
-            fnv1a.HashCore(data, 0, data.Length);
-
-            return fnv1a.fnvhash;
+                HashCore(buffer, 0, encodedLength);
+            }
+            finally
+            {
+                ArrayPool<byte>.Shared.Return(buffer);
+            }
         }
 
         protected override void HashCore(byte[] array, int ibStart, int cbSize)
@@ -95,15 +74,14 @@ namespace CP77.Common.Tools.FNV1A
             var ibEnd = ibStart + cbSize;
             if (ibEnd > array.Length)
             {
-                throw new IndexOutOfRangeException(); 
+                throw new IndexOutOfRangeException();
             }
 
             unchecked
             {
                 for (int i = ibStart; i < ibEnd; i++)
                 {
-                    this.fnvhash ^= array[i];
-                    this.fnvhash *= FnvHashPrime;
+                    this.fnvhash = (this.fnvhash ^ array[i]) * FnvHashPrime;
                 }
             }
         }
@@ -111,6 +89,61 @@ namespace CP77.Common.Tools.FNV1A
         protected override byte[] HashFinal()
         {
             return BitConverter.GetBytes(this.HashUInt64);
+        }
+
+
+        public static ulong HashString(string value)
+        {
+            return HashString(value, Encoding.ASCII, false);
+        }
+
+        public static ulong HashString(string value, Encoding encoding, bool nullEnded = false)
+        {
+            var length = encoding.GetMaxByteCount(nullEnded ? value.Length + 1 : value.Length);
+            var buffer = ArrayPool<byte>.Shared.Rent(length);
+            try
+            {
+                var encodedLength = encoding.GetBytes(value.AsSpan(), buffer.AsSpan());
+                if (nullEnded)
+                {
+                    buffer[encodedLength] = 0;
+                    encodedLength++;
+                }
+
+                return HashReadOnlySpan(new ReadOnlySpan<byte>(buffer, 0, encodedLength));
+            }
+            finally
+            {
+                ArrayPool<byte>.Shared.Return(buffer);
+            }
+        }
+
+        public static ulong HashReadOnlySpan(ReadOnlySpan<char> source)
+        {
+            var hash = FnvHashInitial;
+            foreach (var b in source)
+            {
+                unchecked
+                {
+                    hash = (hash ^ b) * FnvHashPrime;
+                }
+            }
+
+            return hash;
+        }
+
+        public static ulong HashReadOnlySpan(ReadOnlySpan<byte> source)
+        {
+            var hash = FnvHashInitial;
+            foreach (var b in source)
+            {
+                unchecked
+                {
+                    hash = (hash ^ b) * FnvHashPrime;
+                }
+            }
+
+            return hash;
         }
     }
 }