FNV is used in some hot paths such as loading hashes (over 1.5 million times per app start).
This PR improves FNV implementations in two ways:
- Removed allocations by reusing arrays from pool
- Added simplified Span-based versions of algorithm without string re-encoding (hashing is mostly used for filenames with ascii subset)

Benchmark: hashing 1.5m items from `archivehashes.zip`
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-8750H CPU 2.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.101
  [Host]   : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT
  ShortRun : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  

```
|                                          Method |     Mean |    Error |  StdDev | Ratio |       Gen 0 | Gen 1 | Gen 2 |   Allocated |
|------------------------------------------------ |---------:|---------:|--------:|------:|------------:|------:|------:|------------:|
|                         &#39;Original hashing impl&#39; | 446.6 ms | 13.94 ms | 0.76 ms |  1.00 | 105000.0000 |     - |     - | 498528288 B |
|                   &#39;This PR string hashing impl&#39; | 181.8 ms | 38.82 ms | 2.13 ms |  0.41 |           - |     - |     - |       304 B |
| &#39;This PR span-based (no encoding) hashing impl&#39; | 110.4 ms | 12.12 ms | 0.66 ms |  0.25 |           - |     - |     - |       202 B |